Add Vim tab spec lines to files
For most files that is "noexpandtab ts=8 sts=4", which is a de facto in most
files. I find this tabs/spaces policy to be ridiculous - all tabs or all
whitespaces make more sense. But changing formatting post-factum is perhaps
even worse. It seems that this formatting stems from emacs defaults (see
https://github.com/foss-for-synopsys-dwc-arc-processors/toolchain/issues/29#issuecomment-25972243),
it is also used in GCC, but not in bintuils, for example. That should give a
terrible headache for anyone who us using text editor that only has boolean
option to "replace tabs with whitespaces". But this patch at least will
automate things to Vim users.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>