Build ELF32 toolchain without unified source
Back in times all GNU projects lieved in same repository. Then they
splitted but it as assumed that to build stuff one would need to recreate
unified source tree that would mimic old single source tree. There is even a
script that helps to do this with symlinks. That's why it was used in GNU
toolchain build process as well. It seems however that this way of building
doing is not used by many, as I haven't seen any guide or build system to
use it, instead everyone builds binutils, GCC, libc and GDB separately. So
now we also do this. The positive side is that we avoid need to build
symlink source tree. On my rather powerful machine with 6C/12T Xeon and SSD
step for unified source tree building takes around 6 minites. Quite
significant because on baremetal toolchain without multilib takes 3.5 minutes
to build. So removing unified source tree step reduced build time 3x times
in best case. In the worse case of system with multilib build time reduces
from 28 minutes to 22. Numbers would be different for different machines, of
course.

This patch introduces change to baremetal toolchain. Step for unified source
tree is not removed yet, it is becomes useless for toolchain.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>