Fetch even in detached state
Previously arc-versions.sh assumed that it is not possible to fetch in
detached state, so it used to checkout to arbitrary branch before fetching.
That assumption wasn't correct, because it is actually possible to fetch in
detached state, what is actually impossible is to *pull* in this condition.
Thus it is entirely possible to remove this rather redundant check.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>